refactor: Refactor project structure
