+++ src/AppRoutes.js
@@ -1,27 +1,25 @@
-﻿import { Menu } from "./components/Menu";
-import { Cart } from "./components/Cart";
-import { Home } from "./components/Home";
-import { ItemPage } from "./components/ItemDrawer";
-import { Admin } from "./components/Admin";
-
+﻿import { Admin } from './components/Admin/Admin';
+import { Cart } from './components/Cart/Cart';
+import { Home } from './components/Home/Home';
+import { Menu } from './components/Menu/Menu';
 
 const AppRoutes = [
-    {
-        path: "/",
-        element: <Home />
-    },
-    {
-        path: '/menu',
-        element: <Menu />
-    },
-    {
-        path: '/cart',
-        element: <Cart />
-    },
-    {
-        path: '/admin',
-        element: <Admin />
-    },
+  {
+    path: '/',
+    element: <Home />,
+  },
+  {
+    path: '/menu',
+    element: <Menu />,
+  },
+  {
+    path: '/cart',
+    element: <Cart />,
+  },
+  {
+    path: '/admin',
+    element: <Admin />,
+  },
 ];
 
 export default AppRoutes;
