+++ src/components/Home/Home.css
@@ -1,0 +1,102 @@
+body {
+  background-color: #f8f9fa;
+}
+
+.hero-image {
+  background-image: url('https://i.fbcd.co/products/original/img-2589-916de7b33515f0a04eb9d75a93489a699537d20b386b4b547135d913e545a4da.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+  height: 400px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.hero-content {
+  text-align: center;
+}
+
+.menu-item {
+  background-color: #fff;
+  padding: 20px;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition: box-shadow 0.3s ease;
+}
+
+.menu-item:hover {
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.hero-image {
+  background-image: url('https://i.fbcd.co/products/original/img-2589-916de7b33515f0a04eb9d75a93489a699537d20b386b4b547135d913e545a4da.jpg');
+  background-repeat: no-repeat;
+  background-size: cover;
+  background-position: center;
+  height: 400px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.hero-content {
+  text-align: center;
+}
+
+.menu-item {
+  background-color: #fff;
+  padding: 20px;
+  border-radius: 8px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  transition: box-shadow 0.3s ease;
+}
+
+.menu-item:hover {
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.btn-primary {
+  background-color: #ff686b;
+  border-color: #ff686b;
+}
+
+.btn-primary:hover {
+  background-color: #ff5053;
+  border-color: #ff5053;
+}
+
+.btn-secondary {
+  background-color: #ffd86b;
+  border-color: #ffd86b;
+  color: black;
+}
+
+.btn-secondary:hover {
+  background-color: #ffc72d;
+  border-color: #ffc72d;
+}
+
+.footer {
+  background-color: #343a40;
+  color: #fff;
+  padding: 20px;
+  text-align: center;
+}
+
+.footer {
+  background-color: #343a40;
+  color: #fff;
+  padding: 20px;
+  text-align: center;
+}
+
+/* promoting item */
+.promoting-items > .col-md-4 {
+  display: flex;
+  align-items: stretch;
+} 
+
+.mt-5 h2 {
+    margin-bottom: 25px;
+}
