+++ src/components/ItemDrawer/ItemDrawer.css
@@ -1,0 +1,151 @@
+﻿.item-name {
+  font-size: 28px;
+  color: var(--color-pastel-primary);
+  margin-bottom: 10px;
+}
+
+.item-description {
+  font-size: 18px;
+  color: var(--color-pastel-secondary);
+  margin-bottom: 10px;
+}
+
+.item-price {
+  font-size: 20px;
+  color: var(--color-pastel-tertiary);
+  margin-bottom: 20px;
+}
+
+.decorators-section {
+  margin-top: 20px;
+}
+
+.decorators-title {
+  font-size: 22px;
+  color: var(--color-pastel-primary);
+  margin-bottom: 10px;
+}
+
+.decorators-list {
+  list-style-type: none;
+  padding: 0;
+  margin: 0;
+}
+
+.decorator-item {
+  margin-bottom: 10px;
+}
+
+.decorator-checkbox {
+  color: var(--color-pastel-quaternary);
+}
+
+/* Custom checkbox styles */
+.ant-drawer-wrapper {
+  background-color: var(--color-pastel-primary);
+  padding: 20px;
+}
+
+.ant-checkbox-input {
+  width: 26px;
+  height: 26px;
+}
+
+.ant-checkbox-wrapper {
+  color: var(--color-pastel-primary);
+  font-size: 20px;
+}
+
+.ant-checkbox-checked .ant-checkbox-inner,
+.decorator-checkbox .ant-checkbox-input:checked + .ant-checkbox-inner {
+  background-color: var(--color-pastel-primary);
+  border-color: var(--color-pastel-primary);
+  transform: scale(1.3);
+}
+
+.ant-checkbox-checked .ant-checkbox-inner::after,
+.decorator-checkbox .ant-checkbox-input:checked + .ant-checkbox-inner::after {
+  border-color: var(--color-pastel-primary);
+}
+
+/* Media queries for responsiveness */
+@media screen and (max-width: 767px) {
+  .item-name {
+    font-size: 24px;
+  }
+
+  .item-description {
+    font-size: 16px;
+  }
+
+  .item-price {
+    font-size: 18px;
+  }
+
+  .decorators-title {
+    font-size: 18px;
+  }
+
+  .ant-checkbox-input {
+    width: 22px;
+    height: 22px;
+  }
+
+  .ant-checkbox-wrapper {
+    font-size: 18px;
+  }
+}
+
+@media screen and (min-width: 768px) and (max-width: 1023px) {
+  .item-name {
+    font-size: 26px;
+  }
+
+  .item-description {
+    font-size: 17px;
+  }
+
+  .item-price {
+    font-size: 19px;
+  }
+
+  .decorators-title {
+    font-size: 20px;
+  }
+
+  .ant-checkbox-input {
+    width: 24px;
+    height: 24px;
+  }
+
+  .ant-checkbox-wrapper {
+    font-size: 19px;
+  }
+}
+
+@media screen and (min-width: 1024px) {
+  .item-name {
+    font-size: 28px;
+  }
+
+  .item-description {
+    font-size: 18px;
+  }
+
+  .item-price {
+    font-size: 20px;
+  }
+
+  .decorators-title {
+    font-size: 22px;
+  }
+
+  .ant-checkbox-input {
+    width: 26px;
+    height: 26px;
+  }
+
+  .ant-checkbox-wrapper {
+    font-size: 20px;
+  }
+}
